gcc14 int-conversion: record ABI-spec param types; reverse cast at spec-dispatch
gde-generic-dict-eq-map: `__inst_Eq_..._int64_t(a, b)` passes a
`tur_adt_Map__cstr__int *` arg `b` into the spec's int64 param (Class A reverse:
pointer arg -> int64 param). The emit_sig ground-truth table was populated only
from the forward-decl pass over top-level items, so ABI specializations (not
top-level) had no recorded signature -- the call-site reverse cast could not tell
the param was the int64 carrier.

- emit_module.c: record each ABI spec's ACTUAL emitted param C type into emit_sig,
  keyed by its clone name (== the call-site fn_name), in the spec forward-decl
  emitter.
- emit_expr.c: at the dict/spec-dispatch call arg loop, consult emit_sig by
  fn_name; when the recorded param is the int64 carrier and the arg emits as a
  concrete pointer (and is not already bridged), reinterpret through intptr_t.

Fixes gde-generic-dict-eq-map (1->0). This also corrected a latent int-conversion
in van-laarhoven-lens-wide-compose (2 masked errors -> the pointer-into-int64
arg now cast); its codegen snapshot is regenerated. Full suite: 2202 passed,
0 failed.

This completes the carrier-to-typed-param int-conversion front: all 46 flagged
fixtures now compile clean under -Werror=int-conversion.

+++ tests/fixtures/van-laarhoven-lens-wide-compose/expected.c
@@ -4274,7 +4274,7 @@ static int64_t point_hyx_un_undict_un1328(int64_t __dict_1329, int64_t g, int64_
 struct __env_1334 { int64_t __fn; int64_t __hkt_dict; int64_t g; };
 static int64_t __fn_1332(void * __env_p_1335, tur_adt_Point * p) {
         struct __env_1334 *__env___env_1334 = (struct __env_1334 *)__env_p_1335;
-        __auto_type __ps_52 = (point_hyx_un_undict_un1328(__env___env_1334->__hkt_dict, (int64_t)(intptr_t)(__env___env_1334->g), p));
+        __auto_type __ps_52 = (point_hyx_un_undict_un1328(__env___env_1334->__hkt_dict, (int64_t)(intptr_t)(__env___env_1334->g), (int64_t)(intptr_t)(p)));
         if (tur_panicking) return (int64_t){0};
         return __ps_52;
 }
